requirements.txt
Django
stripe
django-environ
